+++ src/components/related-page/__tests__/related-page-test-cases.js
@@ -64,6 +64,44 @@ testCases.example = {
   }
 };
 
+testCases.video = {
+  component: RelatedPage,
+  description: 'video',
+  props: {
+    contentType: 'video',
+    title: 'How to eject a Style Component in Mapbox Studio',
+    children:
+      'Style Components provide sensible defaults and quick opportunities for customization by optimizing the most common property changes for styles and packaging them into simple drop-down options, sliders, and toggles.',
+    vimeoId: '378704089',
+    vimeoThumbnail: 'assets/vimeo_thumb.jpg'
+  }
+};
+
+testCases.videoNoImage = {
+  component: RelatedPage,
+  description: 'video with svg image fallback',
+  props: {
+    contentType: 'video',
+    title: 'How to eject a Style Component in Mapbox Studio',
+    children:
+      'Style Components provide sensible defaults and quick opportunities for customization by optimizing the most common property changes for styles and packaging them into simple drop-down options, sliders, and toggles.',
+    vimeoId: '378704089'
+  }
+};
+
+testCases.videoNoVimeo = {
+  component: RelatedPage,
+  description: 'video with url (no modal)',
+  props: {
+    contentType: 'video',
+    title: 'How to eject a Style Component in Mapbox Studio',
+    children:
+      'Style Components provide sensible defaults and quick opportunities for customization by optimizing the most common property changes for styles and packaging them into simple drop-down options, sliders, and toggles.',
+    url:
+      'https://docs.mapbox.com/android/maps/examples/add-a-vector-tile-source/'
+  }
+};
+
 testCases.children = {
   description: 'fallback with children',
   element: (