+++ src/components/related-page/__tests__/related-page.test.js
@@ -88,6 +88,7 @@ describe('related-page', () => {
       expect(tree).toMatchSnapshot();
     });
   });
+
   describe(testCases.children.description, () => {
     let testCase;
     let wrapper;
@@ -96,6 +97,41 @@ describe('related-page', () => {
     beforeEach(() => {
       testCase = testCases.children;
       wrapper = renderer.create(testCase.element);
+      tree = wrapper.toJSON();
+    });
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe(testCases.video.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = testCases.video;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe(testCases.videoNoImage.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = testCases.videoNoImage;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
       tree = wrapper.toJSON();
     });